Add some price calculator helpers
