+++ apps/store/src/pages/price/[id].tsx
@@ -1,12 +1,8 @@
 import type { GetServerSideProps, NextPage } from 'next'
-import type { FormEvent } from 'react'
+import { setupPriceCalculator } from '@/components/PriceCalculator/PriceCalculator.helpers'
+import { useHandleSubmitPriceCalculator } from '@/components/PriceCalculator/useHandleSubmitPriceCalculator'
 import { TestPricePage } from '@/components/TestPricePage/TestPricePage'
-import { prepopulateFormTemplate } from '@/components/TestPricePage/TestPricePage.helpers'
 import { TestPricePageProps } from '@/components/TestPricePage/TestPricePage.types'
-import useRouterRefresh from '@/hooks/useRouterRefresh'
-import { PageLink } from '@/lib/PageLink'
-import { FormTemplateService } from '@/services/formTemplate/FormTemplateService'
-import { priceIntentServiceInitServerSide } from '@/services/priceIntent/PriceIntentService'
 
 type Params = {
   id: string
@@ -15,24 +11,7 @@ type Params = {
 type Props = TestPricePageProps
 
 const NextTestPricePage: NextPage<Props> = ({ product, template }) => {
-  const refreshData = useRouterRefresh()
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    const formData = new FormData(event.currentTarget)
-    const { intent, ...data } = Object.fromEntries(formData.entries())
-
-    const url = PageLink.apiPriceProduct({
-      productId: product.id,
-      intent: intent === 'confirm' ? 'confirm' : undefined,
-    })
-    await fetch(url, {
-      method: 'POST',
-      body: JSON.stringify(data),
-      headers: { 'Content-Type': 'application/json' },
-    })
-    await refreshData()
-  }
+  const { handleSubmit } = useHandleSubmitPriceCalculator({ productId: product.id })
 
   return (
     <form onSubmit={handleSubmit}>
@@ -43,35 +22,32 @@ const NextTestPricePage: NextPage<Props> = ({ product, template }) => {
 
 export const getServerSideProps: GetServerSideProps<Props, Params> = async (context) => {
   if (context.params === undefined) return { notFound: true }
-
   const productId = context.params.id
 
-  const priceIntentServiceServerSide = priceIntentServiceInitServerSide(context.req, context.res)
+  try {
+    const { template, priceIntent } = await setupPriceCalculator({
+      productId,
+      request: context.req,
+      response: context.res,
+    })
 
-  const formTemplateService = new FormTemplateService()
+    const lineItem = priceIntent.lines?.[0]
 
-  const [emptyTemplate, priceIntent] = await Promise.all([
-    formTemplateService.fetch({ id: productId }),
-    priceIntentServiceServerSide.fetch(productId),
-  ])
-
-  if (emptyTemplate === null) return { notFound: true }
-
-  const template = prepopulateFormTemplate(emptyTemplate, priceIntent.data)
-
-  const lineItem = priceIntent.lines?.[0]
-
-  return {
-    props: {
-      template,
-      product: {
-        id: productId,
-        name: 'Hedvig Home',
-        price: lineItem?.price.amount ?? null,
-        currencyCode: 'SEK',
-        gradient: ['#00BFFF', '#00ff00'],
+    return {
+      props: {
+        template,
+        product: {
+          id: productId,
+          name: 'Hedvig Home',
+          price: lineItem?.price.amount ?? null,
+          currencyCode: 'SEK',
+          gradient: ['#00BFFF', '#00ff00'],
+        },
       },
-    },
+    }
+  } catch (error) {
+    console.error(error)
+    return { notFound: true }
   }
 }
